+++ Engine/Poo.h
@@ -6,19 +6,19 @@
 class Poo
 {
 private:
-	static constexpr int width = 24;
-	static constexpr int height = 24;
-	int x;
-	int y;
-	int velocityOfX;
-	int velocityOfY;
+	static constexpr float width = 24;
+	static constexpr float height = 24;
+	float x;
+	float y;
+	float velocityOfX;
+	float velocityOfY;
 	bool isEaten = false;
 	//bool initialized = false;
 public:
 	void Update();
 	void Colliding(const Player& player);
 	void Draw(Graphics& gfx) const;
-	void Initializer(int xIn, int yIn, int vx, int vy);
+	void Initializer(float xIn, float yIn, float vx, float vy);
 	void GameOver();
 	bool IsEaten() const;
 };
